1.initial README file (Branched out from drm-next-4.6-11-STABLE 8a27ff6)
2.applied x32 patch

3.fixed DDB support
  sys/queue.h: Added DDB_LIST_HEAD to resolve LIST_HEAD conflict
  (kernel vs linuxkpi)

+++ sys/compat/linuxkpi/common/include/linux/interval_tree.h
@@ -8,10 +8,10 @@ struct interval_tree_node {
 		RB_ENTRY(interval_tree_node) rb_entry;
 		struct rb_node rb;
 	};
-	unsigned long start;	/* Start of interval */
-	unsigned long last;	/* Last location _in_ interval */
-	unsigned long __subtree_last;
-	unsigned long it_magic;
+	unsigned long long start;	/* Start of interval */
+	unsigned long long last;	/* Last location _in_ interval */
+	unsigned long long __subtree_last;
+	unsigned long long it_magic;
 };
 
 extern void
@@ -22,10 +22,10 @@ interval_tree_remove(struct interval_tree_node *node, struct rb_root *root);
 
 extern struct interval_tree_node *
 interval_tree_iter_first(struct rb_root *root,
-			 unsigned long start, unsigned long last);
+			 unsigned long long start, unsigned long long last);
 
 extern struct interval_tree_node *
 interval_tree_iter_next(struct interval_tree_node *node,
-			unsigned long start, unsigned long last);
+			unsigned long long start, unsigned long long last);
 
 #endif
